move play button, like, stars below album metadata
